requirements.txt
alpaca-py==0.38.0
yfinance==1.2.0
pandas==2.2.3
ta==0.11.0
vaderSentiment==3.3.2
python-telegram-bot==21.11.1
tenacity==9.0.0
requests==2.32.3
python-dotenv==1.0.1
fredapi==0.5.2
